Refactor Json data handling
