+++ src/InterviewApp/InterviewApp/Server/Data/CsvDataHandler.cs
@@ -1,0 +1,12 @@
+﻿using InterviewApp.Shared.Models;
+
+namespace InterviewApp.Server.Data
+{
+    public class CsvDataHandler : IDataHandler
+    {
+        public List<InterviewQuestion> GetInterviewQuestions()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
