+++ src/InterviewApp/InterviewApp/Server/Data/IDataHandler.cs
@@ -1,0 +1,9 @@
+﻿using InterviewApp.Shared.Models;
+
+namespace InterviewApp.Server.Data
+{
+    public interface IDataHandler
+    {
+        List<InterviewQuestion> GetInterviewQuestions();
+    }
+}
