--- src/InterviewApp/InterviewApp/Shared/CsvDataHandler.cs
@@ -1,13 +1,0 @@
-﻿using InterviewApp.Shared.Interface;
-using InterviewApp.Shared.Models;
-
-namespace InterviewApp.Shared
-{
-    public class CsvDataHandler : IDataHandler
-    {
-        public List<InterviewQuestion> GetDataFromFile()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
