--- src/InterviewApp/InterviewApp/Shared/Interface/IDataHandler.cs
@@ -1,9 +1,0 @@
-﻿using InterviewApp.Shared.Models;
-
-namespace InterviewApp.Shared.Interface
-{
-    public interface IDataHandler
-    {
-        List<InterviewQuestion> GetDataFromFile();
-    }
-}
